+++ woodwork/__init__.py
@@ -2,4 +2,4 @@
 from .data_table import DataTable, DataColumn
 from .version import __version__
 import woodwork.demo
-
+from .config import config